requirements.txt
python-telegram-bot
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
requests
APScheduler
